server/requirements.txt
click==6.7
Flask==1.0.2
peewee==3.3.4
PyMySQL==0.8.1
python-gitlab==1.5.1
Flask-Login==0.4.1
